+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 - Options: `--update-json`, `--skip-images`
 - `--limit`
     - Use the query to limit results. For example, to only download the last week of uploads: `created_at.gt:1 week ago`
+- `--save-json`
+    - Json will only be saved if `--json-path` is given
 
 ## [3.0.0-alpha.1] - 2021-04-21